Desrialization refactor, scene operations (open/save) in editor, double buffering of events

+++ FoolsEngine/src/FoolsEngine/Scene/GameplayWorld/System/SystemsDirector.cpp
@@ -84,7 +84,7 @@ namespace fe
 
 			if (found)
 			{
-				for (int last = enrolls[i].size() - 1; enrollPos < last; ++enrollPos)
+				for (size_t last = enrolls[i].size() - 1; enrollPos < last; ++enrollPos)
 				{
 					std::swap(enrolls[i][enrollPos], enrolls[i][enrollPos + 1]);
 				}
